add generating data to utils
