+++ dirty_cat/tests/test_fast_hash.py
@@ -1,22 +1,5 @@
-import random
-
 from dirty_cat._fast_hash import ngram_min_hash
-
-
-def generate_data(n_samples):
-    MAX_LIMIT = 255  # extended ASCII Character set
-    i = 0
-    str_list = []
-    for i in range(n_samples):
-        random_string = "category "
-        for _ in range(100):
-            random_integer = random.randint(0, MAX_LIMIT)
-            random_string += chr(random_integer)
-            if random_integer < 50:
-                random_string += "  "
-        i += 1
-        str_list += [random_string]
-    return str_list
+from dirty_cat.tests.test_utils import generate_data
 
 
 def test_fast_hash():
